[eas-cli] replace printCredentials with printCredentialsBeta
