+++ packages/eas-cli/src/credentials/ios/actions/SetupBuildCredentials.ts
@@ -10,7 +10,7 @@ import Log from '../../../log';
 import { Action, CredentialsManager } from '../../CredentialsManager';
 import { Context } from '../../context';
 import { AppLookupParams as GraphQLAppLookupParams } from '../api/GraphqlClient';
-import { displayProjectCredentials } from '../utils/printCredentialsBeta';
+import { displayProjectCredentials } from '../utils/printCredentials';
 import { SetupAdhocProvisioningProfile } from './SetupAdhocProvisioningProfile';
 import { SetupInternalProvisioningProfile } from './SetupInternalProvisioningProfile';
 import { SetupProvisioningProfile } from './SetupProvisioningProfile';
